Keep GraphQLTestClient generic
